+++ components/filemngr/listing/api/src/commonMain/kotlin/com/flipperdevices/filemanager/listing/api/FilesDecomposeComponent.kt
@@ -1,6 +1,7 @@
 package com.flipperdevices.filemanager.listing.api
 
 import com.arkivanov.decompose.ComponentContext
+import com.flipperdevices.bridge.connection.feature.storage.api.model.ListingItem
 import com.flipperdevices.ui.decompose.DecomposeOnBackParameter
 import com.flipperdevices.ui.decompose.ScreenDecomposeComponent
 import okio.Path
@@ -8,6 +9,8 @@ import okio.Path
 abstract class FilesDecomposeComponent(
     componentContext: ComponentContext
 ) : ScreenDecomposeComponent(componentContext) {
+    abstract fun onFileChanged(listingItem: ListingItem)
+
     fun interface Factory {
         @Suppress("LongParameterList")
         operator fun invoke(